+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Use Docker Label `com.gaiaadm.pumba.skip=true` to make Pumba ignoring the container. Avoid abusing it.
+
 ## [0.1.6] - 2016-04-25
 ### Added
 - added `gosu` to Pumba Docker image
 ### Fixed
 - Pumba runs as `pumba:pumba` user, instead of `root`
-
 
 ## [0.1.5] - 2016-04-13
 ### Added